+++ src/de/dhbw/Microcontroller/Befehle/Instruction.java
@@ -39,7 +39,7 @@ public class Instruction {
 
     public void incrementProgramCounter(){
         // TODO: Richtig behandeln. Unterschied PC, PCL, PLATH?
-        memory.setAddress(Const.PCL, (byte) (memory.getAddress(Const.PCL) + 1));
+        memory.setAddress(Const.PCL, (memory.getAddress(Const.PCL) + 1));
     }
 
     public void displayDebugInfo(){
@@ -47,7 +47,7 @@ public class Instruction {
     }
 
     public boolean zeroFlagIsSet(){
-        Byte status = memory.getAddress(Const.STATUS);
+        int status = memory.getAddress(Const.STATUS);
 
         if (((status >> 3) & 1) == 1)   return true;
         else    return false;
@@ -55,27 +55,27 @@ public class Instruction {
 
     public void setZeroFlag()
     {
-        byte status = memory.getAddress(Const.STATUS);
-        status = (byte) (status | (1<<2));
+        int status = memory.getAddress(Const.STATUS);
+        status = (status | (1<<2));
         memory.setAddress(Const.STATUS, status);
     }
 
     public void clearZeroFlag()
     {
-        Byte status = memory.getAddress(Const.STATUS);
-        status = (byte) (status & ~(1 << 2));
+        int status = memory.getAddress(Const.STATUS);
+        status = (status & ~(1 << 2));
         memory.setAddress(Const.STATUS, status);
     }
 
     public void setCarryFlag(){
-        byte status = (byte) memory.getAddress(Const.STATUS);
-        status = (byte) (status | (1<<0));
+        int status =  memory.getAddress(Const.STATUS);
+        status = (status | (1<<0));
         memory.setAddress(Const.STATUS, status);
     }
 
     public void clearCarryFlag(){
-        byte status = (byte) memory.getAddress(Const.STATUS);
-        status = (byte) (status & ~(1 << 0));
+        int status = memory.getAddress(Const.STATUS);
+        status = (status & ~(1 << 0));
         memory.setAddress(Const.STATUS, status);
     }
