+++ src/de/dhbw/Microcontroller/Befehle/PIC/CLRF.java
@@ -1,13 +1,12 @@
 package de.dhbw.Microcontroller.Befehle.PIC;
 
+import de.dhbw.Microcontroller.Befehle.Instruction;
+
 /*
  * Clear f
  * TODO: '0x00 -> f; 1 -> Z'
  */
 
-
-import de.dhbw.Constants.Const;
-import de.dhbw.Microcontroller.Befehle.Instruction;
 
 public class CLRF extends Instruction{
     public CLRF(int instruction, int opcode, int argument){
@@ -16,8 +15,8 @@ public class CLRF extends Instruction{
 
     @Override
     public void execute(){
-        byte f = (byte) argument;
-        memory.setAddress(f, (byte) 0);
+        int f = argument;
+        memory.setAddress(f, 0);
 
         setZeroFlag();
