+++ src/de/dhbw/Microcontroller/Befehle/PIC/MOVLW.java
@@ -16,7 +16,7 @@ public class MOVLW extends Instruction {
 
     @Override
     public void execute(){
-        byte k = (byte) argument;
+        int k = argument;
 
         memory.setRegisterW(k);
